ai_features/requirements.txt
fastapi
openai
Pillow
PyMuPDF
python-dotenv
python-multipart
fastapi
uvicorn